src'/IS.ScaleModelsShop.Application/Features/Categories/Commands/CreateCategory/CreateCategoryCommandHandler.cs
using AutoMapper;
using IS.ScaleModelsShop.Application.Repositories;
using IS.ScaleModelsShop.Domain.Entities;
using MediatR;

namespace IS.ScaleModelsShop.Application.Features.Categories.Commands.CreateCategory;

public class CreateCategoryCommandHandler : IRequestHandler<CreateCategoryCommand, CreateCategoryDTO>
{
    private readonly ICategoryRepository _categoryRepository;
    private readonly IMapper _mapper;

    public CreateCategoryCommandHandler(ICategoryRepository categoryRepository, IMapper mapper)
    {
        _categoryRepository = categoryRepository ?? throw new ArgumentNullException(nameof(categoryRepository));
        _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
    }

    public async Task<CreateCategoryDTO> Handle(CreateCategoryCommand request, CancellationToken cancellationToken)
    {
        _ = request ?? throw new ArgumentNullException(nameof(request));

        var category = new Category
        {
            Name = request.Name,
            Id = Guid.NewGuid()
        };

        category = await _categoryRepository.AddAsync(category);

        var createdCategory = _mapper.Map<CreateCategoryDTO>(category);

        return createdCategory;
    }
}